requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pre-commit==3.5.0
black==23.11.0
flake8==6.1.0
pytest==7.4.0
httpx==0.24.0
requests==2.32.0

